+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Implement `color_button` fine-grained setters (e.g. `cg.red = 144`)
 - Support hex colors in `color_button` (by passing an Integer 0xFFBBAA or String hex 'ffbbaa') and `path` `fill` and `stroke` (with `:color` key)
 - Support `path` `transform` property for building `matrix` objects (either `path { transform {operations} }` directly or `path { m1 = matrix {operations}; transform m1 }`)
+- Eliminate need for passing `area_draw_params` to `path` when declared underneath `area` `on_draw` listener
 
 ## 0.1.4